+++ bench-build/CMakeFiles/libzarr_standalone_headers.dir/cmake_clean.cmake
@@ -1,0 +1,22 @@
+file(REMOVE_RECURSE
+  "CMakeFiles/libzarr_standalone_headers.dir/standalone/libzarr_adapters_filesystem_store_hpp.cpp.o"
+  "CMakeFiles/libzarr_standalone_headers.dir/standalone/libzarr_array_hpp.cpp.o"
+  "CMakeFiles/libzarr_standalone_headers.dir/standalone/libzarr_codecs_blosc_hpp.cpp.o"
+  "CMakeFiles/libzarr_standalone_headers.dir/standalone/libzarr_codecs_gzip_hpp.cpp.o"
+  "CMakeFiles/libzarr_standalone_headers.dir/standalone/libzarr_codecs_hpp.cpp.o"
+  "CMakeFiles/libzarr_standalone_headers.dir/standalone/libzarr_codecs_zstd_hpp.cpp.o"
+  "CMakeFiles/libzarr_standalone_headers.dir/standalone/libzarr_group_hpp.cpp.o"
+  "CMakeFiles/libzarr_standalone_headers.dir/standalone/libzarr_libzarr_hpp.cpp.o"
+  "CMakeFiles/libzarr_standalone_headers.dir/standalone/libzarr_metadata_hpp.cpp.o"
+  "CMakeFiles/libzarr_standalone_headers.dir/standalone/libzarr_sharding_hpp.cpp.o"
+  "CMakeFiles/libzarr_standalone_headers.dir/standalone/libzarr_store_hpp.cpp.o"
+  "CMakeFiles/libzarr_standalone_headers.dir/standalone/libzarr_types_hpp.cpp.o"
+  "CMakeFiles/libzarr_standalone_headers.dir/standalone/libzarr_v2_hpp.cpp.o"
+  "CMakeFiles/libzarr_standalone_headers.dir/standalone/libzarr_v3_hpp.cpp.o"
+  "CMakeFiles/libzarr_standalone_headers.dir/standalone/libzarr_zip_hpp.cpp.o"
+)
+
+# Per-language clean rules from dependency scanning.
+foreach(lang CXX)
+  include(CMakeFiles/libzarr_standalone_headers.dir/cmake_clean_${lang}.cmake OPTIONAL)
+endforeach()